First steps towards unit testing
I have added some basic tests to familiarize myself with xunit.
I have also started work on a few basic functions that do not require
the database. After that, I'll begin adding some integration tests
to the mix to make sure the data calls are working properly. The plan
is to use EF Core as the ORM. We'll see how that goes.