Fixed missing FK fields in Area and Square tables
In preparing to move to the next logical tests (EF Core for Area)
I discovered that the FK relationship to Site was missing. Digging
further, I found Square was not related to Area. Doh! I got those
fixed in the creation script, created a new script to generate the
required values in each table as expected by the Integration tests,
and updated the models accordingly.